changed int to float

+++ beehive/Bees.cs
@@ -3,14 +3,14 @@ namespace beehive
     public class Bees
     {
         public string Name;
-        public int Size; //use float for decimal
+        public float Size; //use float for decimal
 
         public Bees()
         {
             this.Name= "";
             this.Size = 0;
         }
-        public Bees(string name, int size)
+        public Bees(string name, float size)
         {
             this.Name = name;
             this.Size = size;
